+++ Arbiter.App/ViewModels/Dialogs/DialogViewModel.cs
@@ -10,7 +10,7 @@ public partial class DialogViewModel : ViewModelBase
     [ObservableProperty] private string? _name;
     [ObservableProperty] private string? _content;
     [ObservableProperty] private int? _sprite;
-    [ObservableProperty] private int? _entityId;
+    [ObservableProperty] private long? _entityId;
     [ObservableProperty] private int? _pursuitId;
     [ObservableProperty] private int? _stepId;