+++ Arbiter.Net/NetworkPriority.cs
@@ -1,0 +1,7 @@
+﻿namespace Arbiter.Net;
+
+public enum NetworkPriority
+{
+    Normal = 0,
+    High = 1,
+}